feat: moved validation into parser
