+++ pixi.toml
@@ -19,3 +19,4 @@ requests = ">=2.32.3,<3"
 bs4 = ">=4.13.4,<5"
 lxml = ">=5.4.0,<6"
 black = ">=25.1.0,<26"
+matplotlib = ">=3.10.3,<4"
