--- src/stats_analysis/README.md
@@ -1,14 +1,0 @@
-# Stats Analysis
-Return statistical test results given counts
-
-## Tests
-- RRR
-- PRR
-- ROR 
-- PhiCorr
-
-All of these rely on the following calculations:
-- Number of events (AE/SE) Y with Drug X
-- Number of events (AE/SE) Y without Drug X
-- Number of events without AE/SE Y with Drug X
-- Number of events without AE/SE Y without Drug X
